feat(`Equipment`): 將`giveEquipment`物件化並整理
