+++ src/core/equipment/table.ts
